# restapi1-src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Base Backend API",
        "description": "Документация для API теста сервера.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/"
        }
    ],
    "paths": {
...
        "/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Получить список комментариев",
                "operationId": "e21e2ae66afe62e7d8d0b023798c70af",
                "parameters": [
                    {
                        "name": "cursor",
                        "in": "query",
                        "description": "Курсор для пагинации",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": ""
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество элементов на странице (по умолчанию 15, максимум 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "commentable_type",
                        "in": "query",
                        "description": "Тип комментируемого объекта для фильтрации",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "App\\\\Models\\\\Product"
                        }
                    },
                    {
                        "name": "commentable_id",
                        "in": "query",
                        "description": "ID комментируемого объекта для фильтрации",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "parent_id",
                        "in": "query",
                        "description": "ID родительского комментария для получения ответов",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": null
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentSectionResource"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Создать комментарий",
                "operationId": "478e0f4964ebbe8d3a22df34f4fcddab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "content": {
                                        "description": "Содержание комментария",
                                        "type": "string",
                                        "example": "Отличный продукт!"
                                    },
                                    "commentable_type": {
                                        "description": "Тип комментируемого объекта",
                                        "type": "string",
                                        "example": "App\\\\Models\\\\Product"
                                    },
                                    "commentable_id": {
                                        "description": "ID комментируемого объекта",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "parent_id": {
                                        "description": "ID родительского комментария (для ответов)",
                                        "type": "integer",
                                        "example": null
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "comment": {
                                            "$ref": "#/components/schemas/CommentResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/comments/{id}": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Получить комментарий по ID",
                "operationId": "c2aac1315d33473f4b25ddf45941b871",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID комментария",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "comment": {
                                            "$ref": "#/components/schemas/CommentResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Комментарий не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "Обновить комментарий",
                "operationId": "0904942451844b2582af15f7c38988b4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID комментария",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "content": {
                                        "description": "Содержание комментария",
                                        "type": "string",
                                        "example": "Обновленный текст комментария"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешно обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "comment": {
                                            "$ref": "#/components/schemas/CommentResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Комментарий не найден"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Удалить комментарий",
                "operationId": "dc71fcedd4e6d8fe9a74a965257cc90b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID комментария",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно удален",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Comment deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Комментарий не найден"
                    }
                }
            }
        },
        "/comments/{id}/replies": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Получить ответы на комментарий",
                "operationId": "46abbc022b3a64df3110d0bdfbee71c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID родительского комментария",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "replies": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CommentResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Комментарий не найден"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "CommentResource": {
                "description": "Комментарий",
                "properties": {
                    "id": {
                        "description": "ID",
                        "type": "integer"
                    },
                    "content": {
                        "description": "Содержание комментария",
                        "type": "string"
                    },
                    "user": {
                        "description": "Автор комментария",
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "name": {
                                "type": "string"
                            },
                            "email": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "commentable_id": {
                        "description": "ID комментируемого объекта",
                        "type": "integer"
                    },
                    "commentable_type": {
                        "description": "Тип комментируемого объекта",
                        "type": "string"
                    },
                    "parent_id": {
                        "description": "ID родительского комментария",
                        "type": "integer",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Дата обновления",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "CommentSectionResource": {
                "description": "Секция комментариев с пагинацией",
                "properties": {
                    "data": {
                        "description": "Список комментариев",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CommentResource"
                        }
                    },
                    "next_cursor": {
                        "description": "Курсор для следующей страницы",
                        "type": "string",
                        "nullable": true
                    },
                    "prev_cursor": {
                        "description": "Курсор для предыдущей страницы",
                        "type": "string",
                        "nullable": true
                    },
                    "per_page": {
                        "description": "Количество элементов на странице",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "security": [
        {
            "accessToken": []
        }
    ],
...
}